organize all hass logic to match frontend project
