+++ src/hass/data/device_registry.ts
@@ -1,0 +1,9 @@
+/**
+ * https://github.com/home-assistant/frontend/blob/dev/src/data/device_registry.ts
+ */
+
+import type { RegistryEntry } from './registry';
+
+export interface DeviceRegistryEntry extends RegistryEntry {
+  area_id: string;
+}
