+++ src/hass/data/entity_registry.ts
@@ -1,0 +1,9 @@
+/**
+ * https://github.com/home-assistant/frontend/blob/dev/src/data/entity_registry.ts
+ */
+
+export interface EntityRegistryDisplayEntry {
+  area_id: string;
+  device_id: string;
+  labels: string[];
+}
