+++ src/hass/data/registry.ts
@@ -1,0 +1,5 @@
+/**
+ * https://github.com/home-assistant/frontend/blob/dev/src/data/registry.ts
+ */
+
+export interface RegistryEntry {}
