--- src/types/editor.ts
@@ -1,5 +1,0 @@
-import type { Config } from '@type/config';
-
-export interface ConfigChangedEvent {
-  config: Config;
-}
