+++ src/util/render.ts
@@ -1,11 +1,11 @@
 import { actionHandler, handleClickAction } from '@common/action-handler';
-import { feature } from '@common/feature';
+import { hasFeature } from '@common/feature';
+import type { HomeAssistant } from '@hass/types';
 import {
   getEntityIconStyles,
   getProblemEntitiesStyle,
 } from '@theme/render-styles';
 import type { Config, EntityInformation } from '@type/config';
-import type { HomeAssistant } from '@type/homeassistant';
 import { html, nothing, type TemplateResult } from 'lit';
 import { getDevice, getEntity, getState } from '../helpers';
 
@@ -17,7 +17,7 @@ export const renderLabel = (
   hass: HomeAssistant,
   config: Config,
 ): TemplateResult | typeof nothing => {
-  if (!hass || feature(config, 'hide_climate_label')) return nothing;
+  if (!hass || hasFeature(config, 'hide_climate_label')) return nothing;
 
   const temp = getState(hass, config.temperature_sensor);
   const humidity = getState(hass, config.humidity_sensor);
@@ -48,7 +48,7 @@ export const renderAreaStatistics = (
   hass: HomeAssistant,
   config: Config,
 ): TemplateResult | typeof nothing => {
-  if (!hass || feature(config, 'hide_area_stats')) return nothing;
+  if (!hass || hasFeature(config, 'hide_area_stats')) return nothing;
 
   const devices = Object.keys(hass.devices).filter(
     (k) => getDevice(hass, k).area_id === config.area,
