+++ test/common/fire-event.spec.ts
@@ -1,5 +1,5 @@
-import { fireEvent } from '@common/fire-event';
-import { type ConfigChangedEvent } from '@type/editor';
+import { fireEvent } from '@hass/common/dom/fire_event';
+import type { ConfigChangedEvent } from '@hass/panels/lovelace/editor/hui-element-editor';
 import { expect } from 'chai';
 import { stub } from 'sinon';
 
@@ -21,6 +21,7 @@ describe('fire-event.ts', () => {
   });
 
   it('should create and dispatch a custom event with detail', () => {
+    // @ts-ignore
     fireEvent(element, 'hass-action', {
       config: { entity: 'light.test' },
       action: 'hold',
@@ -31,6 +32,7 @@ describe('fire-event.ts', () => {
 
     // Retrieve the event argument passed to dispatchEvent
     const event = dispatchStub.firstCall.args[0] as CustomEvent;
+    // @ts-ignore
     expect(event.type).to.equal('hass-action');
     expect(event.bubbles).to.be.true;
     expect(event.composed).to.be.true;
@@ -39,6 +41,7 @@ describe('fire-event.ts', () => {
   });
 
   it('should create and dispatch a custom event with no detail', () => {
+    // @ts-ignore
     fireEvent(element, 'hass-action');
 
     // Ensure the stub was called once
@@ -46,6 +49,7 @@ describe('fire-event.ts', () => {
 
     // Retrieve the event argument passed to dispatchEvent
     const event = dispatchStub.firstCall.args[0] as CustomEvent;
+    // @ts-ignore
     expect(event.type).to.equal('hass-action');
     expect(event.bubbles).to.be.true;
     expect(event.composed).to.be.true;
@@ -57,7 +61,11 @@ describe('fire-event.ts', () => {
       config: { area: 'test-area' },
     };
 
-    const event = fireEvent(element, 'config-changed', configDetail);
+    const event = fireEvent(
+      element,
+      'config-changed',
+      configDetail,
+    ) as CustomEvent;
 
     // Ensure the stub was called once
     expect(dispatchStub.calledOnce).to.be.true;
@@ -66,6 +74,7 @@ describe('fire-event.ts', () => {
   });
 
   it('should work with Window as target', () => {
+    // @ts-ignore
     fireEvent(window, 'hass-action');
 
     // Ensure the stub was called once
@@ -73,6 +82,7 @@ describe('fire-event.ts', () => {
 
     // Retrieve the event argument passed to dispatchEvent
     const event = windowStub.firstCall.args[0] as CustomEvent;
+    // @ts-ignore
     expect(event.type).to.equal('hass-action');
     expect(event.bubbles).to.be.true;
     expect(event.composed).to.be.true;
@@ -80,9 +90,11 @@ describe('fire-event.ts', () => {
   });
 
   it('should return the fired event', () => {
+    // @ts-ignore
     const event = fireEvent(element, 'hass-action');
 
     expect(event).to.be.instanceOf(CustomEvent);
+    // @ts-ignore
     expect(event.type).to.equal('hass-action');
     expect(event.bubbles).to.be.true;
     expect(event.composed).to.be.true;
