+++ test/editor.spec.ts
@@ -1,7 +1,7 @@
 import { RoomSummaryCardEditor } from '@/editor';
+import type { HomeAssistant } from '@hass/types';
 import { fixture } from '@open-wc/testing-helpers';
 import type { Config } from '@type/config';
-import type { HomeAssistant } from '@type/homeassistant';
 import { expect } from 'chai';
 import { nothing, type TemplateResult } from 'lit';
 import { stub } from 'sinon';
